Temporary Access Flow is setup, only issue is start date and end dat not working.
Currently after 10mins of creation access will be granted, and after 10mins access will be revoked.

And terminate session is also not in place yet. It’s not working.

And in Order Guide, restriction for user with access, is not in place.

And Temp access extension is not in place.

And Revoke ServiceNow Admin Access flow is pending.

And in request setup if owner is requester don’t send approval