Temporary Access Flow is setup, only issue is start date and end date not working.
Currently after 10mins of creation access will be granted, and after 10mins access will be revoked.

And in Order Guide, restriction for user with access, is not in place.

And Temp access extension is not in place.

And Revoke ServiceNow Admin Access flow is pending.

And in request setup if owner is requester don’t send approval